% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc2dd.R
\name{char2choice}
\alias{char2choice}
\title{Simple function to generate REDCap choices from character vector}
\usage{
char2choice(data, char.split = "/", raw = NULL, .default = NA)
}
\arguments{
\item{data}{vector}

\item{char.split}{splitting character(s)}

\item{raw}{specific values. Can be used for options of same length.}

\item{.default}{default value for missing. Default is NA.}
}
\value{
vector
}
\description{
Simple function to generate REDCap choices from character vector
}
\examples{
char2choice(c("yes/no","  yep. / nope  ","",NA,"what"),.default=NA)
}
