% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc2dd.R
\name{char2cond}
\alias{char2cond}
\title{Simple function to generate REDCap branching logic from character vector}
\usage{
char2cond(
  data,
  minor.split = ",",
  major.split = ";",
  major.sep = " or ",
  .default = NA
)
}
\arguments{
\item{data}{vector}

\item{minor.split}{minor split}

\item{major.split}{major split}

\item{major.sep}{argument separation. Default is " or ".}

\item{.default}{default value for missing. Default is NA.}
}
\value{
vector
}
\description{
Simple function to generate REDCap branching logic from character vector
}
\examples{
#data <- dd_inst$betingelse
#c("Extubation_novent, 2; Pacu_delay, 1") |> char2cond()
}
