% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_styling.R
\name{create_html_table}
\alias{create_html_table}
\title{Create two-column HTML table for data piping in REDCap instruments}
\usage{
create_html_table(text, variable)
}
\arguments{
\item{text}{descriptive text}

\item{variable}{variable to pipe}
}
\value{
character vector
}
\description{
Create two-column HTML table for data piping in REDCap instruments
}
\examples{
create_html_table(text = "Patient ID", variable = c("[cpr]"))
create_html_table(text = paste("assessor", 1:2, sep = "_"), variable = c("[cpr]"))
# create_html_table(text = c("CPR nummer","Word"), variable = c("[cpr][1]", "[cpr][2]", "[test]"))
}
