% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds2ical.R
\name{ds2ical}
\alias{ds2ical}
\title{Convert data set to ical file}
\usage{
ds2ical(
  data,
  start,
  location,
  summary.glue.string = "ID {id} [{assessor}]",
  description.glue.string = NULL,
  event.length = lubridate::hours(2)
)
}
\arguments{
\item{data}{data set}

\item{start}{event start column}

\item{location}{event location column}

\item{summary.glue.string}{character string to pass to glue::glue() for event
name (summary). Can take any column from data set.}

\item{description.glue.string}{character string to pass to glue::glue() for
event description. Can take any column from data set.}

\item{event.length}{use lubridate functions to generate "Period" class
element (default is lubridate::hours(2))}
}
\value{
tibble of class "ical"
}
\description{
Convert data set to ical file
}
\examples{
df <- dplyr::tibble(start = c(Sys.time(), Sys.time() + lubridate::days(2)),
id = c("1", 3), assessor = "A", location = "111", note = c(NA, "OBS"))
df |> ds2ical(start, location)
df |> ds2ical(start, location,
summary.glue.string = "ID {id} [{assessor}] {note}")
# Export .ics file: (not run)
ical <- df |> ds2ical(start, location, description.glue.string = "{note}")
# ical |> calendar::ic_write(file=here::here("calendar.ics"))
}
