% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easy_redcap.R
\name{easy_redcap}
\alias{easy_redcap}
\title{Secure API key storage and data acquisition in one}
\usage{
easy_redcap(project.name, widen.data = TRUE, uri, ...)
}
\arguments{
\item{project.name}{The name of the current project (for key storage with
`keyring::key_set()`, using the default keyring)}

\item{widen.data}{argument to widen the exported data}

\item{uri}{REDCap database API uri}

\item{...}{arguments passed on to `REDCapCAST::read_redcap_tables()`}
}
\value{
data.frame or list depending on widen.data
}
\description{
Secure API key storage and data acquisition in one
}
