% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{focused_metadata}
\alias{focused_metadata}
\title{focused_metadata}
\usage{
focused_metadata(metadata, vars_in_data)
}
\arguments{
\item{metadata}{A dataframe containing metadata}

\item{vars_in_data}{Vector of variable names in the dataset}
}
\value{
A dataframe containing metadata for the variables in the dataset
}
\description{
Extracts limited metadata for variables in a dataset
}
