% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_styling.R
\name{format_subheader}
\alias{format_subheader}
\title{Sub-header formatting wrapper}
\usage{
format_subheader(data, tag = "h2")
}
\arguments{
\item{data}{character vector}

\item{tag}{character vector length 1}
}
\value{
character vector
}
\description{
Sub-header formatting wrapper
}
\examples{
"Instrument header" |> format_subheader()
}
