% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easy_redcap.R
\name{get_api_key}
\alias{get_api_key}
\title{Retrieve project API key if stored, if not, set and retrieve}
\usage{
get_api_key(key.name)
}
\arguments{
\item{key.name}{character vector of key name}
}
\value{
character vector
}
\description{
Retrieve project API key if stored, if not, set and retrieve
}
