% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds2dd_detailed.R
\name{hms2character}
\alias{hms2character}
\title{Change "hms" to "character" for REDCap upload.}
\usage{
hms2character(data)
}
\arguments{
\item{data}{data set}
}
\value{
data.frame or tibble
}
\description{
Change "hms" to "character" for REDCap upload.
}
\examples{
data <- redcapcast_data
## data |> time_only_correction() |> hms2character()
}
