% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_styling.R
\name{html_tag_wrap}
\alias{html_tag_wrap}
\title{Simple html tag wrapping for REDCap text formatting}
\usage{
html_tag_wrap(data, tag = "h2", extra = NULL)
}
\arguments{
\item{data}{character vector}

\item{tag}{character vector length 1}

\item{extra}{character vector}
}
\value{
character vector
}
\description{
Simple html tag wrapping for REDCap text formatting
}
\examples{
html_tag_wrap("Titel", tag = "div", extra = 'class="rich-text-field-label"')
html_tag_wrap("Titel", tag = "h2")
}
