% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{is.repeated_longitudinal}
\alias{is.repeated_longitudinal}
\title{Test if repeatable or longitudinal}
\usage{
is.repeated_longitudinal(
  data,
  generics = c("redcap_event_name", "redcap_repeat_instrument", "redcap_repeat_instance")
)
}
\arguments{
\item{data}{data set}

\item{generics}{default is "redcap_event_name", "redcap_repeat_instrument"
and "redcap_repeat_instance"}
}
\value{
logical
}
\description{
Test if repeatable or longitudinal
}
\examples{
is.repeated_longitudinal(c("record_id", "age", "record_id", "gender"))
is.repeated_longitudinal(redcapcast_data)
is.repeated_longitudinal(list(redcapcast_data))
}
