% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc2dd.R
\name{is_missing}
\alias{is_missing}
\title{Multi missing check}
\usage{
is_missing(data, nas = c("", "NA"))
}
\arguments{
\item{data}{character vector}

\item{nas}{character vector of strings considered as NA}
}
\value{
logical vector
}
\description{
Multi missing check
}
