% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds2dd_detailed.R
\name{mark_complete}
\alias{mark_complete}
\title{Completion marking based on completed upload}
\usage{
mark_complete(upload, ls)
}
\arguments{
\item{upload}{output list from `REDCapR::redcap_write()`}

\item{ls}{output list from `ds2dd_detailed()`}
}
\value{
list with `REDCapR::redcap_write()` results
}
\description{
Completion marking based on completed upload
}
