% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{match_fields_to_form}
\alias{match_fields_to_form}
\title{Match fields to forms}
\usage{
match_fields_to_form(metadata, vars_in_data)
}
\arguments{
\item{metadata}{A data frame containing field names and form names}

\item{vars_in_data}{A character vector of variable names}
}
\value{
A data frame containing field names and form names
}
\description{
Match fields to forms
}
