% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtcars_redcap.R
\docType{data}
\name{mtcars_redcap}
\alias{mtcars_redcap}
\title{mtcars dataset slightly modified to use for Shiny app upload demonstration}
\format{
A data frame with 13 variables:
\describe{
  \item{record_id}{ID, numeric}
  \item{mpg}{ID, numeric}
  \item{cyl}{ID, numeric}
  \item{disp}{ID, numeric}
  \item{hp}{ID, numeric}
  \item{drat}{ID, numeric}
  \item{wt}{ID, numeric}
  \item{qsec}{ID, numeric}
  \item{vs}{ID, numeric}
  \item{am}{ID, numeric}
  \item{gear}{ID, numeric}
  \item{carb}{ID, numeric}
  \item{name}{original rownames, charater}

}
}
\usage{
data(mtcars_redcap)
}
\description{
mtcars dataset slightly modified to use for Shiny app upload demonstration
}
\keyword{datasets}
