% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_user_input.r
\name{process_user_input.character}
\alias{process_user_input.character}
\title{User input processing character}
\usage{
\method{process_user_input}{character}(x, ...)
}
\arguments{
\item{x}{input}

\item{...}{ignored}
}
\value{
processed input
}
\description{
User input processing character
}
