% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_user_input.r
\name{process_user_input.data.frame}
\alias{process_user_input.data.frame}
\title{User input processing data.frame}
\usage{
\method{process_user_input}{data.frame}(x, ...)
}
\arguments{
\item{x}{input}

\item{...}{ignored}
}
\value{
processed input
}
\description{
User input processing data.frame
}
