% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_user_input.r
\name{process_user_input}
\alias{process_user_input}
\title{User input processing}
\usage{
process_user_input(x)
}
\arguments{
\item{x}{input}
}
\value{
processed input
}
\description{
User input processing
}
