% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_user_input.r
\name{process_user_input.response}
\alias{process_user_input.response}
\title{User input processing response}
\usage{
\method{process_user_input}{response}(x, ...)
}
\arguments{
\item{x}{input}

\item{...}{ignored}
}
\value{
processed input
}
\description{
User input processing response
}
