% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_redcap_tables.R
\name{read_redcap_tables}
\alias{read_redcap_tables}
\title{Download REDCap data}
\usage{
read_redcap_tables(
  uri,
  token,
  records = NULL,
  fields = NULL,
  events = NULL,
  forms = NULL,
  raw_or_label = "label",
  generics = c("record_id", "redcap_event_name", "redcap_repeat_instrument",
    "redcap_repeat_instance"),
  ...
)
}
\arguments{
\item{uri}{REDCap database uri}

\item{token}{API token}

\item{records}{records to download}

\item{fields}{fields to download}

\item{events}{events to download}

\item{forms}{forms to download}

\item{raw_or_label}{raw or label tags}

\item{generics}{vector of auto-generated generic variable names to
ignore when discarding empty rows}

\item{...}{ekstra parameters for REDCapR::redcap_read_oneshot}
}
\value{
list of instruments
}
\description{
Wrapper function for using REDCapR::redcap_read and REDCapRITS::REDCap_split
including some clean-up. Works with longitudinal projects with repeating
instruments.
}
\examples{
# Examples will be provided later
}
