% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{sanitize_split}
\alias{sanitize_split}
\title{Sanitize list of data frames}
\usage{
sanitize_split(
  l,
  generic.names = c("redcap_event_name", "redcap_repeat_instrument",
    "redcap_repeat_instance")
)
}
\arguments{
\item{l}{A list of data frames.}

\item{generic.names}{A vector of generic names to be excluded.}
}
\value{
A list of data frames with generic names excluded.
}
\description{
Removing empty rows
}
