% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_cast.R
\name{shiny_cast}
\alias{shiny_cast}
\title{Launch the included Shiny-app for database casting and upload}
\usage{
shiny_cast()
}
\value{
shiny app
}
\description{
Launch the included Shiny-app for database casting and upload
}
\examples{
# shiny_cast()

}
