% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds2dd_detailed.R
\name{time_only_correction}
\alias{time_only_correction}
\title{Correction based on time_only_filter function}
\usage{
time_only_correction(data, ...)
}
\arguments{
\item{data}{data set}

\item{...}{arguments passed on to `guess_time_only_filter()`}
}
\value{
tibble
}
\description{
Correction based on time_only_filter function
}
\examples{
data <- redcapcast_data
## data |> time_only_correction()
}
